[trel] Add cli to get TREL UDP port.
The test software for certification currently uses the mDNS packets of
trel service responses to find out the trel port, and then use the port
number to determine which packets in a capture should be decoded as TREL
packets. However this may not be reliable since it depends on when the
capture starts. Added a cli to get trel port, so this can be used by
a THCI function.

(cherry picked from commit 68544b0709b8c337fe605fe71a5877ee5e3d2a5d)
Change-Id: I07b9952732a0feacf8580df9e4ba4dbcad6acfb7